requirements.txt
psycopg2-binary
allure-python-commons
pytest
requests
selenium
allure-pytest